Refresh button refactor
