+++ src/components/Main/Main.jsx
@@ -4,7 +4,7 @@ import { useParams } from 'react-router-dom';
 
 import BreadcrumbContainer from '../Breadcrumb';
 import FilesContainer from '../Files';
-import RefreshButtonContainer from '../RefreshButton';
+import RefreshButtonContainer from '../refresh-button';
 
 import './style/style.scss';
