--- src/components/RefreshButton/index.js
@@ -1,1 +1,0 @@
-export { default } from './RefreshButton.container';
