+++ src/components/refresh-button/index.js
@@ -1,0 +1,1 @@
+export { default } from './refresh-button.container';
