+++ src/components/refresh-button/refresh-button.container.js
@@ -4,7 +4,7 @@ import { useParams } from 'react-router-dom';
 
 import { fetchFileListFromPath } from 'store/actions/filesActions';
 
-import RefreshButtonComponent from './RefreshButton.component';
+import RefreshButtonComponent from './refresh-button.component';
 
 const RefreshButtonContainer = () => {
   const { path } = useParams();
