+++ src/styles/colors.scss
@@ -6,7 +6,7 @@ $gray: #808080;
 $soft-gray: #F2F3F4;
 $not-so-soft-gray: #808080;
 $icon-gray: #D2DAE2;
-$orange: #FFA834;
+$orange: #ffa834;
 $playing-song: #f0c47e;
 $selected-item: #FFF6EC;
 $broken-link-background: #f0a27e;
